+++ packages/modals/package.json
@@ -40,8 +40,8 @@
     ]
   },
   "dependencies": {
-    "@floating-ui/react": "^0.25.4",
-    "@titicaca/core-elements": "workspace:*"
+    "@titicaca/core-elements": "workspace:*",
+    "@titicaca/tds-ui": "workspace:*"
   },
   "devDependencies": {
     "@titicaca/i18n": "workspace:*",
