+++ packages/modals/src/index.ts
@@ -1,5 +1,6 @@
-export * from './alert'
-export * from './confirm'
-export * from './modal'
+export * from '@titicaca/tds-ui/src/components/alert'
+export * from '@titicaca/tds-ui/src/components/confirm'
+export * from '@titicaca/tds-ui/src/components/modal'
+
 export * from './login-cta-modal'
 export * from './transition-modal'
